requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
SQLAlchemy==2.0.48
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.26
pydantic==2.12.5